Render and visualize Map Services from ArcGIS Online and ArcGIS Server.

Map Services are a way to expose the contents of a map as a web service and expose capabilities for exporting tile images, querying and identifying features and more. Also supports custom popups and identification of features.

addEsriDynamicMapLayer(map, url, options = dynamicMapLayerOptions(),
  popupFunction = NULL, popupOptions = NULL, layerId = NULL,
  group = NULL)

Arguments

map

The leaflet map

url

URL of the http://resources.arcgis.com/en/help/arcgis-rest-api/#/Map_Service/02r3000000w2000000/.

options

options for the dynamic map layer.

popupFunction

Uses the provided function to create a popup that will identify features whenever the map is clicked. Your function will be passed a GeoJSON FeatureCollection of the features at the clicked location and should return the appropriate HTML. If you do not want to open the popup when there are no results, return false.

popupOptions

See popupOptions.

layerId

A unique ID for the layer.

group

The name of the group this layer should be added to.