Options for dynamic map layer.

Options for dynamic map layer.

dynamicMapLayerOptions(format = "png24", transparent = TRUE, f = "json",
  attribution = "", layers = NULL, layerDefs = NULL, opacity = 1,
  position = "front", maxZoom = NULL, minZoom = NULL,
  dynamicLayers = NULL, token = NULL, proxy = NULL, useCors = TRUE, ...)

Arguments

format

Output format of the image.

transparent

Allow the server to produce transparent images.

f

Server response content type.

attribution

Attribution from service metadata copyright text is automatically displayed in Leaflet's default control. This property can be used for customization.

layers

An array of Layer IDs like [3,4,5] to show from the service.

layerDefs

A string representing a query to run against the service before the image is rendered. This can be a string like "3:STATE_NAME='Kansas'" or an object mapping different queries to specific layers 3:"STATE_NAME='Kansas'", 2:"POP2007>25000".

opacity

Opacity of the layer. Should be a value between 0 (completely transparent) and 1 (completely opaque).

position

Position of the layer relative to other overlays.

maxZoom

Closest zoom level the layer will be displayed on the map.

minZoom

Furthest zoom level the layer will be displayed on the map.

dynamicLayers

JSON object literal used to manipulate the layer symbology defined in the service itself. Requires a 10.1 (or above) map service which supports dynamicLayers requests.

token

If you pass a token in your options it will be included in all requests to the service.

proxy

URL of an https://developers.arcgis.com/javascript/jshelp/ags_proxy.html or https://github.com/Esri/resource-proxy to use for proxying requests.

useCors

If this service should use CORS when making GET requests.

...

extra options