Options for featureLayers.

Options for featureLayers.

featureLayerOptions(where = NULL, minZoom = NULL, maxZoom = NULL,
  cacheLayers = NULL, fields = NULL, from = NULL, to = NULL,
  timeField = NULL, timeFilterMode = NULL, simplifyFactor = NULL,
  precision = NULL, token = NULL, proxy = NULL, useCors = NULL,
  renderer = NULL, ...)

Arguments

where

String An optional expression to filter features server side. String values should be denoted using single quotes ie: where: "FIELDNAME = 'field value'"; More information about valid SQL syntax can be found at http://resources.arcgis.com/en/help/main/10.2/index.html#/SQL_reference_for_query_expressions_used_in_ArcGIS/00s500000033000000/.

minZoom

Integer Minimum zoom level of the map that features will display. example: minZoom:0

maxZoom

Integer Maximum zoom level of the map that features will example: maxZoom:19

cacheLayers

Boolean Will remove layers from the internal cache when they are removed from the map.

fields

Array An array of fieldnames to pull from the service. Includes all fields by default. You should always specify the name of the unique id for the service. Usually either 'FID' or 'OBJECTID'.

from

Date When paired with to defines the time range of features to display. Requires the Feature Layer to be time enabled.

to

Date When paired with from defines the time range of features to display. Requires the Feature Layer to be time enabled.

timeField

false The name of the field to lookup the time of the feature. Can be an object like start:'startTime', end:'endTime' or a string like 'created'.

timeFilterMode

'server' (default) or 'client' Determines where features are filtered by time. By default features will be filtered by the server. If set to 'client' all features are requested and filtered by the app before display.

simplifyFactor

Integer How much to simplify polygons and polylines. More means better performance, and less means more accurate representation.

precision

Integer How many digits of precision to request from the server. Wikipedia has a great reference of digit precision to meters at http://en.wikipedia.org/wiki/Decimal_degrees.

token

String If you pass a token in your options it will be included in all requests to the service.

proxy

URL of an https://developers.arcgis.com/javascript/jshelp/ags_proxy.html or https://github.com/Esri/resource-proxy to use for proxying requests.

useCors

Boolean If this service should use CORS when making GET requests.

renderer

L.svg or L.canvas The vector renderer to use to draw the service. Usually L.svg but setting to L.canvas contains performance benefits for large polygon layers.

...

extra options