Options for image map layer.

Options for image map layer.

imageMapLayerOptions(format = "jpgpng", f = "json", opacity = 1,
  position = "front", maxZoom = NULL, minZoom = NULL, from = NULL,
  to = NULL, bandIds = NULL, noData = NULL, noDataInterpretation = NULL,
  pixelType = NULL, renderingRule = NULL, mosaicRule = NULL,
  token = NULL, proxy = NULL, useCors = TRUE, ...)

Arguments

format

Output format of the image.

f

Server response content type.

opacity

Opacity of the layer. Should be a value between 0 and 1.

position

Position of the layer relative to other overlays.

maxZoom

Closest zoom level the layer will be displayed on the map.

minZoom

Furthest zoom level the layer will be displayed on the map.

from

Date When paired with to defines the time range of data to display. Requires the Image Layer to be time enabled.

to

Date When paired with from defines the time range of data to display. Requires the Image Layer to be time enabled.

bandIds

If there are multiple bands, you can specify which bands to export.

noData

The pixel value representing no information.

noDataInterpretation

Interpretation of the noData setting.

pixelType

Leave pixelType as unspecified, or UNKNOWN, in most exportImage use cases, unless such pixelType is desired. Possible values: C128, C64, F32, F64, S16, S32, S8, U1, U16, U2, U32, U4, U8, UNKNOWN.

renderingRule

A JSON representation of a http://resources.arcgis.com/en/help/arcgis-rest-api/#/Raster_function_objects/02r3000000rv000000/

mosaicRule

A JSON representation of a http://resources.arcgis.com/en/help/arcgis-rest-api/#/Mosaic_rule_objects/02r3000000s4000000/

token

If you pass a token in your options it will be included in all requests to the service.

proxy

URL of an https://developers.arcgis.com/javascript/jshelp/ags_proxy.html or https://github.com/Esri/resource-proxy to use for proxying requests.

useCors

If this service should use CORS when making GET requests.

...

extra options